requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
Pillow>=11.0.0
pydantic>=2.10.0
python-multipart>=0.0.20
